test(proxy): 添加 wsrelay 和 errors 单元测试
- manager.go: 修复 ConnectionState 类型转换问题
- executor.go: 移除未使用的 uuid 导入
- 新增 message_test.go: 消息创建测试
- 新增 session_test.go: 会话管理测试（幂等 Close、并发安全）
- 新增 errors_test.go: 错误类型测试（Unwrap、IsRetryableError）

所有测试通过：proxy 包 40+ 测试，wsrelay 包 7 测试，errors 9 测试

+++ proxy/wsrelay/manager.go
@@ -62,7 +62,7 @@ func NewWsConnection(conn *websocket.Conn, session *Session, wsURL string) *WsCo
 		URL:     wsURL,
 	}
 	wc.lastUsed.Store(time.Now().UnixNano())
-	wc.state.Store(StateConnected)
+	wc.state.Store(int32(StateConnected))
 	return wc
 }
 
@@ -79,13 +79,13 @@ func (wc *WsConnection) IsExpired() bool {
 
 // IsConnected 检查是否已连接
 func (wc *WsConnection) IsConnected() bool {
-	return wc.state.Load() == StateConnected
+	return wc.state.Load() == int32(StateConnected)
 }
 
 // Close 安全关闭连接
 func (wc *WsConnection) Close() error {
-	if wc.state.CompareAndSwap(StateConnected, StateClosing) ||
-		wc.state.CompareAndSwap(StateConnecting, StateClosing) {
+	if wc.state.CompareAndSwap(int32(StateConnected), int32(StateClosing)) ||
+		wc.state.CompareAndSwap(int32(StateConnecting), int32(StateClosing)) {
 		// 调用断开回调
 		if wc.onDisconnected != nil && wc.session != nil {
 			wc.onDisconnected(wc.session.AccountID)
